Implemented Graph data type, began defining Prim

+++ EdgeWeightedDigraph/Edge.cpp
@@ -1,6 +1,6 @@
 #include "Edge.h"
 
-Edge::Edge(Vertex* tail, Vertex* head, int weight) {
+Edge::Edge(Vertex* tail, Vertex* head, float weight) {
 	this->tail = tail;
 	this->head = head;
 	this->weight = weight;
